Implement `Lox::Interpreter`
The acceptance tests say:

```
FAIL test/scanning/numbers.lox
     Expected return code 0 and got 1. Stderr:
     lib/lox/cli.rb:9:in `initialize': uninitialized constant Lox::CLI::Interpreter (NameError)

           self.interpreter = Interpreter.new(output, logger)
                              ^^^^^^^^^^^
     	from bin/lox:6:in `new'
     	from bin/lox:6:in `<main>'
```

So we need to implement `Lox::Interpreter`.

Contrary to the chapter, I’ve made the decision that the scanner should
return an `Enumerator` of tokens instead of an array, with support for
externally iterating them with a block passed to `#each_token`. This
feels more idiomatic for Ruby and also gives me the opportunity to make
the scanner work by iterating through the underlying characters and
incrementally yielding tokens as they’re recognised, which is arguably
more interesting.