Start the Lox CLI from the `bin/lox` executable
The acceptance tests say:

```
FAIL test/scanning/numbers.lox
     Missing expected output 'NUMBER 123 123.0' on line 6.
     Missing expected output 'NUMBER 123.456 123.456' on line 7.
     Missing expected output 'DOT . null' on line 8.
     Missing expected output 'NUMBER 456 456.0' on line 9.
     Missing expected output 'NUMBER 123 123.0' on line 10.
     Missing expected output 'DOT . null' on line 11.
     Missing expected output 'EOF  null' on line 12.
```

So we need to make `bin/lox` actually do something.

It’s a single line of glue code to wire up the environment (and the
acceptance tests) to the actual implementation, so there’s no real value
in writing unit tests for it.