Introduce `Lox::Scanner` with support for scanning the empty string
The acceptance tests say:

```
FAIL test/scanning/numbers.lox
     Expected return code 0 and got 1. Stderr:
     lib/lox/interpreter.rb:9:in `run': uninitialized constant Lox::Interpreter::Scanner (NameError)

           scanner = Scanner.new(source, logger)
                     ^^^^^^^
     	from lib/lox/cli.rb:36:in `block in run_file'
     	from lib/lox/cli.rb:35:in `open'
     	from lib/lox/cli.rb:35:in `run_file'
     	from lib/lox/cli.rb:18:in `start'
     	from bin/lox:6:in `<main>'
```

So we need to implement `Lox::Scanner`.

This commit gives us the entire public API of the scanner without any of
the actual guts.

As explained in an earlier commit, the intention is to iterate over the
characters of `source` and incrementally yield them from `#each_token`.
The assumption is that `source` supports `#each_char`, so both strings
and files (or any other variety of `IO` object) will work.

The acceptance tests [0] check the scanner’s features in a different
order than the chapter [1] introduces them. For the sake of clarity
we’re going to follow the chapter order while implementing the scanner;
the acceptance tests should all pass when we’re done.

[0] https://github.com/munificent/craftinginterpreters/tree/master/test/scanning
[1] https://craftinginterpreters.com/scanning.html