Add support for scanning slashes and comments
While the book tries to treat these together, I’ve decided to keep them
distinct by adding a separate stage which skips comments before even
trying to produce a token.

The downside is that we require two-character lookahead to recognise the
start of a comment (“//”), but we’re going to need that later for digits
anyway, so the organisational benefit of separating the comments concern
doesn’t come at any real cost.