Implement lookahead iterator for peeking further into an `Enumerator`
The scanner consumes an `Enumerator` of characters and is therefore only
able to `#peek` [0] a single character ahead. But for digits (and,
optionally, comments) it needs to be able to peek _two_ characters
ahead, e.g. so it can distinguish “42.1” (a number lexeme) from “42.” (a
number lexeme followed by a dot lexeme).

This commit introduces `Lox::LookaheadIterator`, which wraps an
`Enumerator` and exposes the ability to externally iterate it (i.e.
call `#next`) while enriching `#peek` with an optional `lookahead:`
argument. Arbitrary lookahead is supported by iterating the underlying
`Enumerator` and filling a buffer to answer future queries.

Next we’ll use this in the scanner to wrap the `Enumerator` of
characters and thereby make it easy to peek further ahead when we need
to.

[0] https://docs.ruby-lang.org/en/3.1/Enumerator.html#method-i-peek